+++ CHANGELOG.md
@@ -2,6 +2,7 @@ UNRELEASED
 ----------
 
 * [ [#478](https://github.com/digitalfabrik/lunes-cms/issues/478) ] Add possibility to set example sentence
+* [ [#490](https://github.com/digitalfabrik/lunes-cms/issues/490) ] Add module and vocabulary counters to discipline list
 
 
 2023.5.0