+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 UNRELEASED
 ----------
 
+
+2023.7.0
+--------
+
 * [ [#478](https://github.com/digitalfabrik/lunes-cms/issues/478) ] Add possibility to set example sentence
 * [ [#490](https://github.com/digitalfabrik/lunes-cms/issues/490) ] Add module and vocabulary counters to discipline list
 * [ [#490](https://github.com/digitalfabrik/lunes-cms/issues/490) ] Add vocabulary counters to training set list